refactor(dot11): extract label helpers into labels module
